+++ src/integration/agents/src/main/java/se/tink/backend/aggregation/agents/nxgen/serviceproviders/banks/nordea/partner/fetcher/transactional/entity/InterestIntervalsEntity.java
@@ -1,15 +1,16 @@
 package se.tink.backend.aggregation.agents.nxgen.serviceproviders.banks.nordea.partner.fetcher.transactional.entity;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.math.BigDecimal;
 import se.tink.backend.aggregation.annotations.JsonObject;
 
 @JsonObject
 public class InterestIntervalsEntity {
     @JsonProperty("from_amount")
-    private double fromAmount;
+    private BigDecimal fromAmount;
 
     @JsonProperty("to_amount")
-    private double toAmount;
+    private BigDecimal toAmount;
 
-    private double rate;
+    private BigDecimal rate;
 }
