+++ src/integration/agents/src/main/java/se/tink/backend/aggregation/agents/nxgen/serviceproviders/banks/nordea/partner/fetcher/transactional/entity/InterestsEntity.java
@@ -1,6 +1,7 @@
 package se.tink.backend.aggregation.agents.nxgen.serviceproviders.banks.nordea.partner.fetcher.transactional.entity;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.math.BigDecimal;
 import java.util.List;
 import se.tink.backend.aggregation.annotations.JsonObject;
 
@@ -10,13 +11,13 @@ public class InterestsEntity {
     private String rateType;
 
     @JsonProperty("accumulated_amount")
-    private double accumulatedAmount;
+    private BigDecimal accumulatedAmount;
 
     @JsonProperty("accumulated_year_amount")
-    private double accumulatedYearAmount;
+    private BigDecimal accumulatedYearAmount;
 
     @JsonProperty("accumulated_last_year_amount")
-    private double accumulatedLastYearAmount;
+    private BigDecimal accumulatedLastYearAmount;
 
     @JsonProperty("interest_intervals")
     private List<InterestIntervalsEntity> interestIntervals;
